requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
